delete slides, and insert slide title created and fixed the bug

+++ AAUG.DomainModels/ViewModels/SldieShowTitleViewModels.cs
@@ -7,7 +7,7 @@ public class SlideShowTitleInsertViewModel
 
 public class SlideShowTitleGetViewModel
 {
-    public short? Id { get; set; }
+    public int? Id { get; set; }
     public string? Description { get; set; }
     public List<SlideShowGetViewModel> SlideShowGetViewModels { get; set; }
 }
